Add user dto
