+++ backend/Facebook/src/main/java/org/example/Facebook/model/dto/UserDto.java
@@ -1,0 +1,20 @@
+package org.example.Facebook.model.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.Date;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class UserDto {
+    private Integer Id;
+    private String name;
+    private Date birthday;
+    private String email;
+    private String bio;
+}
