+++ backend/Facebook/src/main/java/org/example/Facebook/model/entity/Ad.java
@@ -1,21 +1,21 @@
-package org.example.Facebook.model;
+package org.example.Facebook.model.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
-import java.util.List;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @Entity
-@Table(name = "ad_profiles")
-public class AdProfile {
+@Table(name = "ads")
+public class Ad {
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   @Id
-  private Integer Id;
-  private Integer userId;
-  private String keywords;
+  public String Id;
+  public String title;
+  public String content;
+  public Integer numberOfClicks;
 }
