+++ backend/Facebook/src/main/java/org/example/Facebook/model/entity/Post.java
@@ -1,4 +1,4 @@
-package org.example.Facebook.model;
+package org.example.Facebook.model.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
