+++ backend/Facebook/src/main/java/org/example/Facebook/service/AdManagerService.java
@@ -1,9 +1,9 @@
 package org.example.Facebook.service;
 
 
-import org.example.Facebook.model.Ad;
-import org.example.Facebook.model.AdProfile;
-import org.example.Facebook.model.User;
+import org.example.Facebook.model.entity.Ad;
+import org.example.Facebook.model.entity.AdProfile;
+import org.example.Facebook.model.entity.User;
 import org.springframework.stereotype.Service;
 
 @Service
