+++ backend/Facebook/src/main/java/org/example/Facebook/service/AdminService.java
@@ -1,7 +1,7 @@
 package org.example.Facebook.service;
 
-import org.example.Facebook.model.Post;
-import org.example.Facebook.model.User;
+import org.example.Facebook.model.entity.Post;
+import org.example.Facebook.model.entity.User;
 import org.springframework.stereotype.Service;
 
 @Service
